requirements.txt
Flask>=3.0
Flask-Mail>=0.10
python-dotenv>=1.0
gunicorn>=21.2.0
